プレイヤーあたり改良

+++ Tsuyu/TY/Math.h
@@ -50,7 +50,7 @@ namespace TY
             return value;
         }
 
-        constexpr double Clamp(float value, float min, float max)
+        constexpr float Clamp(float value, float min, float max)
         {
             if (value < min) return min;
             if (value > max) return max;
